+++ src/models/PanelSettings.ts
@@ -1,4 +1,5 @@
 import { FileType } from "vscode";
+import { Choice } from "./Choice";
 import { DashboardData } from "./DashboardData";
 
 export interface PanelSettings {
@@ -27,8 +28,9 @@ export interface Field {
   title?: string;
   name: string;
   type: "string" | "number" | "datetime" | "boolean" | "image" | "choice" | "tags" | "categories";
-  choices?: string[];
+  choices?: string[] | Choice[];
   single?: boolean;
+  multiSelect?: boolean;
 }
 
 export interface DateInfo {